Moved Org model to org app and modified the imports
